Changed to controlled component
