+++ src/Accordion/README.md
@@ -6,33 +6,54 @@ Accordion component implemented (mostly) according to [WAI-ARIA specification](h
 
 ```jsx
 import * as React from "react"
-import { Accordion, AccordionSection } from "@operational/components"
+import { Accordion, useAccordionState } from "@operational/components"
 
-const MyComponent = () => (
-  <div style={{ height: 400 }}>
-    <Accordion>
-      <AccordionSection title="Section 1" expanded>
-        Content 1<br />
-        Content 1<br />
-        Content 1<br />
-        Content 1<br />
-        Content 1<br />
-        Content 1<br />
-        <a href="#">test link</a>
-        <br />
-        Content 1<br />
-        Content 1<br />
-        Content 1<br />
-        Content 1<br />
-        Content 1<br />
-        Content 1<br />
-        Content 1<br />
-      </AccordionSection>
-      <AccordionSection title="Section 2">Content 2</AccordionSection>
-      <AccordionSection title="Section 3">Content 3</AccordionSection>
-    </Accordion>
-  </div>
-)
+const MyComponent = () => {
+  const [sections, onToggle] = useAccordionState([
+    {
+      key: 1,
+      title: "Section 1",
+      expanded: true,
+      content: () => (
+        <>
+          Content 1<br />
+          Content 1<br />
+          Content 1<br />
+          Content 1<br />
+          Content 1<br />
+          Content 1<br />
+          <a href="#">test link</a>
+          <br />
+          Content 1<br />
+          Content 1<br />
+          Content 1<br />
+          Content 1<br />
+          Content 1<br />
+          Content 1<br />
+          Content 1<br />
+        </>
+      ),
+    },
+    {
+      key: 2,
+      title: "Section 2",
+      expanded: true,
+      content: () => <>Content 2</>,
+    },
+    {
+      key: 3,
+      title: "Section 3",
+      expanded: true,
+      content: () => <>Content 3</>,
+    },
+  ])
+
+  return (
+    <div style={{ height: 400 }}>
+      <Accordion sections={sections} onToggle={onToggle} />
+    </div>
+  )
+}
 
 ;<MyComponent />
 ```
