--- src/AccordionSection/README.md
@@ -1,1 +1,0 @@
-AccordionSections are elements used exclusively inside `Accordion` elements. See [Accordion](/#/Components/Accordion) for usage examples.
