+++ src/index.ts
@@ -75,7 +75,7 @@ export { default as Tree, TreeProps } from "./Tree/Tree"
 export { default as TourModal, TourModalProps } from "./TourModal/TourModal"
 export { default as LabelText } from "./LabelText/LabelText"
 export { default as Accordion, AccordionProps } from "./Accordion/Accordion"
-export { default as AccordionSection, AccordionSectionProps } from "./AccordionSection/AccordionSection"
+export { useAccordionState } from "./Accordion/useAccordionState"
 
 // Typography components
 export { default as FinePrint } from "./Typography/FinePrint"
